+++ lib/ringcentral_sdk/cache.rb
@@ -1,3 +1,0 @@
-module RingCentralSdk::Cache
-  autoload :Extensions, 'ringcentral_sdk/cache/extensions'
-end