+++ lib/ringcentral_sdk/helpers.rb
@@ -1,7 +1,0 @@
-module RingCentralSdk::Helpers
-  autoload :CreateFaxRequest, 'ringcentral_sdk/helpers/fax'
-  autoload :ExtensionPresence, 'ringcentral_sdk/helpers/extension_presence'
-  autoload :Inflator, 'ringcentral_sdk/helpers/inflator'
-  autoload :Messages, 'ringcentral_sdk/helpers/messages'
-  autoload :Request, 'ringcentral_sdk/helpers/request'
-end