+++ lib/ringcentral_sdk/helpers/inflator.rb
@@ -1,3 +1,0 @@
-module RingCentralSdk::Helpers::Inflator
-  autoload :ContactInfo, 'ringcentral_sdk/helpers/inflator/contact_info'
-end