+++ lib/ringcentral_sdk/rest.rb
@@ -1,0 +1,9 @@
+module RingCentralSdk::REST
+  autoload :Cache, 'ringcentral_sdk/rest/cache'
+  autoload :Client, 'ringcentral_sdk/rest/client'
+  autoload :Messages, 'ringcentral_sdk/rest/messages'
+  autoload :MessagesFax, 'ringcentral_sdk/rest/messages'
+  autoload :MessagesSMS, 'ringcentral_sdk/rest/messages'
+  autoload :Request, 'ringcentral_sdk/rest/request'
+  autoload :Subscription, 'ringcentral_sdk/rest/subscription'
+end