+++ lib/ringcentral_sdk/rest/cache.rb
@@ -1,0 +1,3 @@
+module RingCentralSdk::REST::Cache
+  autoload :Extensions, 'ringcentral_sdk/rest/cache/extensions'
+end