+++ lib/ringcentral_sdk/rest/extension_presence.rb
@@ -1,16 +1,16 @@
-module RingCentralSdk::Helpers
+module RingCentralSdk
   class ExtensionPresence
 
-    attr_accessor :rc_api
+    attr_accessor :client
     attr_accessor :account_id
-    attr_accessor :exension_id
-    attr_accessor :presence_info
+    attr_accessor :extension_id
+    attr_accessor :presence_data
 
-    def initialize(rc_api, extension_id=nil)
-      @rc_api = rc_api
+    def initialize(extension_id, opts={})
+      @client = opts.has_key?(:client) ? opts[:client] : nil
       @account_id = '~'
       @extension_id = extension_id.to_s
-      @presence_info = {}
+      @presence_data = {}
     end
 
     def retrieve()
@@ -18,23 +18,23 @@ module RingCentralSdk::Helpers
         raise "extension_id is not an integer"
       end
 
-      res = @rc_api.client.get do |req|
+      res = @client.http.get do |req|
         req.url "account/#{@account_id}/extension/#{@extension_id}/presence"
       end
 
-      @presence_info = res.body
+      @presence_data = res.body
 
-      return @presence_info
+      return @presence_data
     end
 
     def department_calls_enable(enable)
       retrieve()
 
-      if !@presence_info.has_key?('dndStatus')
+      if !@presence_data.has_key?('dndStatus')
         raise 'invalid presence info'
       end
 
-      current_status = @presence_info['dndStatus']
+      current_status = @presence_data['dndStatus']
       new_status = enable ?
         status_enable_dnd_department_calls(current_status) :
         status_disable_dnd_department_calls(current_status)
@@ -47,11 +47,11 @@ module RingCentralSdk::Helpers
     def department_calls_enabled?(reload=false)
       if reload
         retrieve()
-      elsif !@presence_info.has_key?('dndStatus')
+      elsif !@presence_data.has_key?('dndStatus')
         retrieve()
       end
 
-      current_status = @presence_info['dndStatus']
+      current_status = @presence_data['dndStatus']
 
       status_enabled = {
         'DoNotAcceptAnyCalls' => false,
@@ -67,11 +67,11 @@ module RingCentralSdk::Helpers
     def disable_department_calls()
       retrieve()
 
-      if !@presence_info.has_key?('dndStatus')
+      if !@presence_data.has_key?('dndStatus')
         raise 'invalid presence info'
       end
 
-      current_status = @presence_info['dndStatus']
+      current_status = @presence_data['dndStatus']
       new_status = status_disable_dnd_department_calls(current_status)
 
       if current_status != new_status
@@ -84,15 +84,15 @@ module RingCentralSdk::Helpers
         raise 'HTTP request body is required to update presence'
       end
 
-      res = @rc_api.client.put do |req|
+      res = @client.http.put do |req|
         req.url "account/#{@account_id}/extension/#{@extension_id}/presence"
         req.headers['Content-Type'] = 'application/json'
         req.body = body
       end
       
-      @presence_info = res.body
+      @presence_data = res.body
 
-      return @presence_info
+      return @presence_data
     end
 
     def status_enable_dnd_department_calls(current_status)