+++ lib/ringcentral_sdk/rest/request.rb
@@ -1,0 +1,5 @@
+module RingCentralSdk::REST::Request
+  autoload :Base, 'ringcentral_sdk/rest/request/base'
+  autoload :Fax, 'ringcentral_sdk/rest/request/fax'
+  autoload :Inflator, 'ringcentral_sdk/rest/request/inflator'
+end