+++ lib/ringcentral_sdk/rest/request/inflator.rb
@@ -1,0 +1,3 @@
+module RingCentralSdk::REST::Request::Inflator
+  autoload :ContactInfo, 'ringcentral_sdk/rest/request/inflator/contact_info'
+end