+++ lib/ringcentral_sdk/rest/request/inflator/contact_info.rb
@@ -1,4 +1,4 @@
-module RingCentralSdk::Helpers::Inflator
+module RingCentralSdk::REST::Request::Inflator
   class ContactInfo
     def inflate_to_array(any=nil)
       contacts   = []