+++ test/test_helper.rb
@@ -2,7 +2,7 @@ require 'coveralls'
 Coveralls.wear!
 
 require 'test/unit'
-require "mocha/test_unit"
+require 'mocha/test_unit'
 require 'ringcentral_sdk'
 
 rcsdk = RingCentralSdk.new(